+++ src/index.ts
@@ -12,7 +12,7 @@ import { editCommand } from "./commands/edit.js";
 import { todayCommand } from "./commands/today.js";
 
 const program = new Command();
-program.name("habitxt").description("Text-based habit tracker").version("1.1.2");
+program.name("habitxt").description("Text-based habit tracker").version("1.1.3");
 
 doCommand(program);
 showCommand(program);